ENH: fix converted package bootstrapping issues when converting setup.py
Because toydist is installed with setuptools, and because setuptools screws up
sys.path, it breaks some bootstrapping issues when executing the setup.py to
convert. In particular, many packages import themselves in setup.py to read the
version, and previously, the version would be the installed package one (if
any) instead of the package to be converted.