ENH: implement file hook + example.
File hook can be used in the case where the hook needs to be
filename-specific (instead of extension-specific). The current way of
doing it is hackish - should find a way to remove the need to
pickle/unpickle hooks, as it forces the hook to be a top-name function
(cannot be a closure for example).