ENH: make decorator callable to deal with help bypass.
One may not want to execute hook if the command has been called in help mode.
Although this can be dealt inside the hook itself, it is easier to deal with
this in the decorator. This breaks all hook files because the decorator needs
to be called now (@foo vs @foo()).