REF: remove task signature code.
All things considered, it is easy to get task signature system wrong , and not
that useful, so we remove it. For most time-consuming phases (build, install),
common dependencies are outside bento control. For example, one would need very
tight integration with the build system to detect changed source files, etc...