ENH+BUG: rewrite lexer from near-scratch.
This lexer is both simpler and more performant. It uses ply states
feature to deal with multilines fields. In particular, the description
field value is lexed as a single token, which should make the parser
much faster. Comments handling should be more reliable as well.

The lexer could most likely be simplified further (to be done later),
and the parser has not been updated to this new lexer yet.